Enhance authentication system and improve app stability
## Major Improvements:

### Authentication & Security
- Add centralized AuthenticationManager with automatic token refresh and re-login
- Implement bearer token authentication for all file upload/download APIs
- Add SmsSubmissionWorker with proper authentication and error handling
- Create UserSession entity for secure token management

### Database & Migration
- Fix AppLaunchStatus nullable return type to prevent migration crashes
- Add UserAccount entity with foreign key constraints
- Update database schema to version 49 with proper migrations
- Improve database error handling in ViewModels

### API & Network
- Add comprehensive subscription and payment API endpoints
- Implement AuthenticationManager.executeWithAuth pattern across all authenticated APIs
- Add file upload/download authentication to BackupWorker and BackupRestoreScreenViewModel
- Create new financial message and subscription response models

### UI & User Experience
- Fix BackupWorker notification auto-dismissal after 3 seconds
- Improve SMS permission handling with graceful fallbacks
- Complete SubscriptionScreenViewModel.getSubscriptionContainer function
- Add proper error handling and loading states

### SMS & Worker Management
- Create SmsSubmissionWorker with smart SIM detection and carrier mapping
- Add comprehensive SMS processing with authentication
- Improve SMS provider detection and phone number handling
- Add proper retry mechanisms and exponential backoff

### Bug Fixes
- Fix race conditions in database initialization during fresh installs
- Resolve SQLiteException crashes with proper Flow error handling
- Fix notification persistence issues in background workers
- Improve permission request flow in SMS fetch screen

+++ app/src/main/java/com/records/pesa/mapper/TransactionsMapper.kt
@@ -55,7 +55,7 @@ fun TransactionItem.toTransaction(userId: Int): Transaction = Transaction(
     comment = comment,
     balance = balance,
     entity = entity,
-    userId = userId
+    userId = userId.toLong()
 )
 
 fun TransactionWithCategories.toTransaction(userId: Int): Transaction = Transaction(
@@ -72,7 +72,7 @@ fun TransactionWithCategories.toTransaction(userId: Int): Transaction = Transact
     comment = transaction.comment,
     balance = transaction.balance,
     entity = transaction.entity,
-    userId = userId
+    userId = userId.toLong()
 )
 
 fun Transaction.toTransactionItem(): TransactionItem = TransactionItem(
